Renames file to be inline with open source package
