+++ datagrowth/resources/http/__init__.py
@@ -1,2 +1,2 @@
 from .generic import HttpResource
-from .downloads import HttpFileResource
+from .files import HttpFileResource
